2025-10-16 - Finalizando o login, register e integrações de banco

+++ SignalR_API/SignalR_Services/User/UserService.cs
@@ -3,6 +3,8 @@ using SignalR_Domains.Interface;
 using SignalR_Domains.User;
 using SignalR_Domains.Users;
 using SignalR_Errors;
+using System;
+using System.Text.RegularExpressions;
 
 namespace SignalR_Services.User
 {
@@ -23,8 +25,10 @@ namespace SignalR_Services.User
         private Guid GetUserIdFromToken()
         {
             var userId = _tokenService.GetUserIdFromToken();
+
             if (userId == Guid.Empty)
-                throw new ErrorLists(new List<string> { "Token inválido." });
+                throw new ErrorLists(["Token inválido."]);
+
             return userId;
         }
 
@@ -35,10 +39,8 @@ namespace SignalR_Services.User
 
             var user = await _userRepository.GetUserByIdAsync(userId!.Value);
 
-            user = new SignalR_Domains.User.User();
-
             if (user == null)
-                throw new ErrorLists(new List<string> { "Usuario não encontrado." });
+                throw new ErrorLists(["Usuario não encontrado."]);
 
             return new UserParams(user);
         }
@@ -54,18 +56,36 @@ namespace SignalR_Services.User
 
         public async Task<List<UserParams>> SearchUsersAsync(string query)
         {
+            if (string.IsNullOrWhiteSpace(query))
+                throw new ErrorLists(["Query inválida."]);   
+
+            query = query.Trim();
+
+            var querys = Regex.Split(query, @"[^A-Za-z0-9À-ÿ]+");
+
             var userId = GetUserIdFromToken();
 
-            var users = await _userRepository.GetUsersByFilterAsync(query);
+            var users = await _userRepository.GetUsersByFilterAsync(querys);
 
             return users.Select(u => new UserParams(u)).ToList();
+        }
+
+        public async Task<bool> HasEmailRegisterAsync(string email)
+        {
+            if (await _userRepository.HasEmailRegisterAsync(email))
+                return true;
+
+            return false;
         }
 
         public async Task<UserParams> CreateUserAsync(UserParams @params)
         {
             var user = await SignalR_Domains.User.User.Create(@params);
 
-            var token = _tokenService.GenerateToken(user);
+            if (await HasEmailRegisterAsync(@params.Email!))
+                throw new ErrorLists(["Email já cadastrado "]);
+
+            await _userRepository.CreateUserAsync(user);
 
             return new UserParams(user);
         }
@@ -76,7 +96,15 @@ namespace SignalR_Services.User
 
             var existingUser = await _userRepository.GetUserByIdAsync(userId);
 
+            if(existingUser.Email != @params.Email)
+            {
+                if (await HasEmailRegisterAsync(@params.Email!))
+                    throw new ErrorLists(["Email já cadastrado "]);
+            }
+
             var updatedUser = await SignalR_Domains.User.User.Update(existingUser, @params);
+        
+            await _userRepository.UpdateUserAsync(updatedUser);
 
             return new UserParams(updatedUser);
         }
@@ -84,6 +112,11 @@ namespace SignalR_Services.User
         public async Task<bool> DeleteUserAsync()
         {
             var userId = _tokenService.GetUserIdFromToken();
+
+            var existingUser = await _userRepository.GetUserByIdAsync(userId);
+
+            if (existingUser == null)
+                throw new ErrorLists(["Usuario não encontrado."]);
 
             return await _userRepository.DeleteUserAsync(userId);
         }
